IKASAN-932 - WIP DB consumer/producer

+++ ikasaneip/developer/mvn-archetype/ikasan-jboss-6-db-im/src/main/resources/archetype-resources/jar/src/main/java/example/db/messageprovider/DBMessageProvider.java
@@ -79,7 +79,11 @@ public class DBMessageProvider implements MessageProvider<Collection<Model>>
     public Collection<Model> invoke(JobExecutionContext context)
     {
         List<Model> entities = ioService.findAllEntities();
-        ioService.remove(entities);
+        if(entities != null)
+        {
+            ioService.remove(entities);
+        }
+
         return entities;
     }
